make Days field decimal in LeaveEntitlementComponent.cs

+++ Ergon.API/Models/LeaveEntitlementComponent.cs
@@ -3,13 +3,11 @@ namespace Ergon.Models
     public class LeaveEntitlementComponent
     {
         public int LeaveEntitlementComponentId { get; set; }
-        public int TotalDays { get; set; }
+        public decimal TotalDays { get; set; }
 
-        // foreign keys
         public int LeaveEntitlementId { get; set; }
         public int LeaveTypeId { get; set; }
 
-        // navigation
         public LeaveEntitlement LeaveEntitlement { get; set; } = null!;
         public LeaveType LeaveType { get; set; } = null!;
     }
